implemented cmds to sequencer proof client cli
