+++ crates/sequencer_proof_client/Cargo.toml
@@ -13,6 +13,7 @@ categories.workspace = true
 [dependencies]
 zkos_wrapper.workspace = true
 zksync_airbender_execution_utils.workspace = true
+zksync_airbender_cli.workspace = true
 bellman.workspace = true
 circuit_definitions.workspace = true
 crypto_codegen.workspace = true
@@ -27,3 +28,6 @@ serde_json.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
+
+[features]
+gpu = ["zksync_airbender_cli/gpu"]
