+++ crates/zksync_os_fri_prover/Cargo.toml
@@ -13,6 +13,7 @@ categories.workspace = true
 [dependencies]
 zksync_airbender_cli.workspace = true
 zksync_airbender_execution_utils.workspace = true
+zksync_sequencer_proof_client = { path = "../sequencer_proof_client" }
 
 anyhow.workspace = true
 base64.workspace = true
