+++ crates/zksync_os_fri_prover/src/lib.rs
@@ -3,20 +3,17 @@ use std::{
     time::SystemTime,
 };
 
-use anyhow::{anyhow, Result};
 use base64::{engine::general_purpose::STANDARD, Engine as _};
 
 use clap::Parser;
-use reqwest::{Client, StatusCode};
-use serde::{Deserialize, Serialize};
 use zksync_airbender_cli::{
     prover_utils::{
-        create_proofs_internal, create_recursion_proofs, load_binary_from_path,
-        program_proof_from_proof_list_and_metadata, GpuSharedState,
+        create_proofs_internal, create_recursion_proofs, load_binary_from_path, program_proof_from_proof_list_and_metadata, serialize_to_file, GpuSharedState
     },
     Machine,
 };
 use zksync_airbender_execution_utils::ProgramProof;
+use zksync_sequencer_proof_client::SequencerProofClient;
 
 /// Command-line arguments for the Zksync OS prover
 #[derive(Parser, Debug)]
@@ -36,75 +33,6 @@ pub struct Args {
     /// Circuit limit - max number of MainVM circuits to instantiate to run the block fully
     #[arg(long, default_value = "10000")]
     pub circuit_limit: usize,
-}
-
-// Note: copied from zkos_prover_input_generator.rs
-#[derive(Debug, Serialize, Deserialize)]
-struct NextProverJobPayload {
-    block_number: u32,
-    prover_input: String, // base64-encoded
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ProofPayload {
-    block_number: u32,
-    proof: String,
-}
-/// HTTP client for the proof-data server
-#[derive(Clone)]
-pub struct ProofDataClient {
-    http: Client,
-    base_url: String,
-}
-
-impl ProofDataClient {
-    /// Create a new client targeting the given base URL (e.g., "http://localhost:3000")
-    pub fn new<U: Into<String>>(base_url: U) -> Self {
-        ProofDataClient {
-            http: Client::new(),
-            base_url: base_url.into(),
-        }
-    }
-
-    /// Fetch the next block to prove.
-    /// Returns `Ok(None)` if there's no block pending (204 No Content).
-    pub async fn pick_next_prover_job(&self) -> Result<Option<(u32, Vec<u8>)>> {
-        let url = format!("{}/prover-jobs/FRI/pick", self.base_url);
-        let resp = self.http.post(&url).send().await?;
-
-        match resp.status() {
-            StatusCode::OK => {
-                let body: NextProverJobPayload = resp.json().await?;
-                let data = STANDARD
-                    .decode(&body.prover_input)
-                    .map_err(|e| anyhow!("Failed to decode block data: {}", e))?;
-                Ok(Some((body.block_number, data)))
-            }
-            StatusCode::NO_CONTENT => Ok(None),
-            s => Err(anyhow!("Unexpected status {} when fetching next block", s)),
-        }
-    }
-
-    /// Submit a proof for the processed block
-    /// Returns the vector of u32 as returned by the server.
-    pub async fn submit_proof(&self, block_number: u32, proof: String) -> Result<()> {
-        let url = format!("{}/prover-jobs/FRI/submit", self.base_url);
-        let payload = ProofPayload {
-            block_number,
-            proof,
-        };
-
-        let resp = self.http.post(&url).json(&payload).send().await?;
-
-        if resp.status().is_success() {
-            Ok(())
-        } else {
-            Err(anyhow!(
-                "Server returned {} when submitting proof",
-                resp.status()
-            ))
-        }
-    }
 }
 
 fn create_proof(
@@ -146,7 +74,7 @@ pub async fn run(args: Args) {
         args.enabled_logging
     );
 
-    let client = ProofDataClient::new(args.base_url);
+    let client = SequencerProofClient::new(args.base_url);
 
     let manifest_path = if let Ok(manifest_path) = std::env::var("CARGO_MANIFEST_DIR") {
         manifest_path
@@ -159,10 +87,10 @@ pub async fn run(args: Args) {
     let binary = load_binary_from_path(&binary_path.to_str().unwrap().to_string());
     let mut gpu_state = GpuSharedState::new(&binary);
 
-    println!("Starting Zksync OS FRI prover for {}", client.base_url);
+    println!("Starting Zksync OS FRI prover for {}", client.sequencer_url());
 
     loop {
-        let (block_number, prover_input) = match client.pick_next_prover_job().await {
+        let (block_number, prover_input) = match client.pick_fri_job().await {
             Err(err) => {
                 eprintln!("Error fetching next prover job: {err}");
                 tokio::time::sleep(tokio::time::Duration::from_millis(1000)).await;
@@ -201,7 +129,10 @@ pub async fn run(args: Args) {
         // 2) base64-encode that binary blob
         let proof_b64 = STANDARD.encode(&proof_bytes);
 
-        match client.submit_proof(block_number, proof_b64).await {
+        let path = format!("./fri_proof.json");
+        serialize_to_file(&proof_b64, Path::new(&path));
+
+        match client.submit_fri_proof(block_number, proof_b64).await {
             Ok(_) => println!(
                 "{:?} successfully submitted proof for block number {}",
                 SystemTime::now(),
